Version 2019-12-03b 5mmb_classic_120319b
This new version is the above version, but it permanently eliminates an
issue with opening windows known as the Unused bug.

Thanks to @Aabacus for the brilliant solution and implementation!

CONSIDER THIS BETA--since window opening timings are changed now. Let me
know if your windows fail to log completely in (I still don't chose your
license) or any other issues.